requirements.txt
pip
pytest
automonisaur
selenium>=3.141.0
beautifulsoup4>=4.10.0
requests>=2.26.0
